+++ server/builder/file-group.ts
@@ -1,4 +1,9 @@
-import { B3File, Config, RawContent } from "./models/interfaces";
+import {
+  B3File,
+  Config,
+  OutputFileTypes,
+  RawContent,
+} from "./models/interfaces";
 import * as path from "path";
 import { tableOfContentsHtml, tableOfContentsMd } from "./ui";
 
@@ -9,16 +14,21 @@ export class FileGroup {
   config: Config = {
     sourceRootPath: "",
     htmlOutputPath: "",
-    outputType: "html",
+    outputType: OutputFileTypes.HTML,
     markdownOutputPath: "",
+    targetCategory: "",
   };
   constructor(config: Config, rawContent: RawContent[]) {
     this.rawContent = rawContent;
     this.config = config;
-    if (this.config.outputType === "md") {
+    if (!this.config.outputType) {
+      // TODO: Move it to logger class
+      throw new Error("Critical Error: Output type should be defined");
+    }
+    if (this.config.outputType === OutputFileTypes.MD) {
       this.outputPath = this.config.markdownOutputPath;
     }
-    if (this.config.outputType === "html") {
+    if (this.config.outputType === OutputFileTypes.HTML) {
       this.outputPath = this.config.htmlOutputPath;
     }
   }
@@ -36,8 +46,10 @@ export class FileGroup {
   }
 
   generateTableOfContents(text: string): string {
-    if (this.config.outputType === "md") return tableOfContentsMd(text);
-    if (this.config.outputType === "html") return tableOfContentsHtml(text);
+    if (this.config.outputType === OutputFileTypes.MD)
+      return tableOfContentsMd(text);
+    if (this.config.outputType === OutputFileTypes.HTML)
+      return tableOfContentsHtml(text);
     return "";
   }
 
@@ -50,11 +62,13 @@ export class FileGroup {
       ),
       content: rawContent.content,
       category: rawContent.category,
-      name: rawContent.metadata.name || "",
+      name: rawContent.metadata.title || "",
+      sort: rawContent.metadata.sort || 0,
+      ignore: rawContent.metadata.ignore || false,
     };
   }
 
-  createAggregatedFileGroup() {
+  createAggregatedFileGroup(groupName: string) {
     const files: B3File[] = [];
     const contentAggregationFromMap = Object.fromEntries(
       this.aggregatedContent
@@ -62,12 +76,21 @@ export class FileGroup {
     Object.keys(contentAggregationFromMap).forEach((key: string) => {
       files.push({
         path:
-          key === "all"
-            ? path.join(this.outputPath, "all." + this.config.outputType)
-            : path.join(this.outputPath, key, "all." + this.config.outputType),
+          key === groupName
+            ? path.join(
+                this.outputPath,
+                groupName + "." + this.config.outputType
+              )
+            : path.join(
+                this.outputPath,
+                key,
+                groupName + "." + this.config.outputType
+              ),
         content: contentAggregationFromMap[key],
         category: key,
-        name: "all",
+        name: groupName,
+        sort: 0,
+        ignore: false,
       });
     });
     return files;
@@ -75,7 +98,11 @@ export class FileGroup {
 
   async run(): Promise<any[]> {
     const files: B3File[] = [];
-
+    if (this.config.targetCategory) {
+      this.rawContent = this.rawContent.filter(
+        (item: RawContent) => item.category === this.config.targetCategory
+      );
+    }
     for (let i = 0; i < this.rawContent.length; i++) {
       const rawContent: RawContent = this.rawContent[i];
 
@@ -96,7 +123,38 @@ export class FileGroup {
       this.generateTableOfContents(this.aggregatedContent.get("all") || "") +
         this.aggregatedContent.get("all")
     );
-    const aggregatedFiles = this.createAggregatedFileGroup();
+    const aggregatedFiles = this.createAggregatedFileGroup("all");
+    files.push(...aggregatedFiles);
+    return files;
+  }
+
+  async prepareBookTemplateContent(): Promise<any[]> {
+    const files: B3File[] = [];
+    if (this.config.targetCategory) {
+      this.rawContent = this.rawContent.filter(
+        (item: RawContent) => item.category === this.config.targetCategory
+      );
+    }
+    for (let i = 0; i < this.rawContent.length; i++) {
+      const rawContent: RawContent = this.rawContent[i];
+
+      this.initAggregatedContentKey("prepared-book-template");
+
+      this.appendAggregatedContentValue(
+        "prepared-book-template",
+        rawContent.content
+      );
+    }
+
+    this.aggregatedContent.set(
+      "prepared-book-template",
+      this.generateTableOfContents(
+        this.aggregatedContent.get("prepared-book-template") || ""
+      ) + this.aggregatedContent.get("prepared-book-template")
+    );
+    const aggregatedFiles = this.createAggregatedFileGroup(
+      "prepared-book-template"
+    );
     files.push(...aggregatedFiles);
     return files;
   }